fix: correct aegir settings & type failures
